+++ tests/parametric/test_span_sampling.py
@@ -557,6 +557,7 @@ class Test_Span_Sampling:
     @missing_feature(context.library == "nodejs", reason="Not implemented")
     @missing_feature(context.library == "php", reason="The PHP tracer always sends the full trace to the agent.")
     @missing_feature(context.library == "python", reason="RPC issue causing test to hang")
+    @missing_feature(context.library == "python_http", reason="RPC issue causing test to hang")
     @missing_feature(
         context.library == "ruby", reason="Issue: sending the complete trace when only the root span is expected"
     )
@@ -613,6 +614,7 @@ class Test_Span_Sampling:
     @missing_feature(context.library == "nodejs", reason="Not implemented")
     @missing_feature(context.library == "php", reason="The PHP tracer always sends the full trace to the agent.")
     @missing_feature(context.library == "python", reason="RPC issue causing test to hang")
+    @missing_feature(context.library == "python_http", reason="RPC issue causing test to hang")
     @missing_feature(
         context.library == "ruby", reason="Issue: sending the complete trace when only the root span is expected"
     )
@@ -670,6 +672,7 @@ class Test_Span_Sampling:
     @missing_feature(context.library == "nodejs", reason="Not implemented")
     @missing_feature(context.library == "php", reason="The PHP tracer always sends the full trace to the agent.")
     @missing_feature(context.library == "python", reason="RPC issue causing test to hang")
+    @missing_feature(context.library == "python_http", reason="RPC issue causing test to hang")
     @missing_feature(
         context.library == "ruby", reason="Issue: sending the complete trace when only the root span is expected"
     )
